refactor common methods to a mixin class and reorganize
